% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_authentication.R
\name{mod_authentification_server}
\alias{mod_authentification_server}
\title{authentification Server Functions}
\usage{
mod_authentification_server(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}
}
\value{
A named list with various reactive values
\describe{
 \item{connect}{A function handling the connection to an Opensilex WebAPI}
 \item{user}{The currently connected user}
 \item{password}{The current password}
 \item{host}{The current opensilex WebAPI host address}
}
}
\description{
The server part of the module handling the connection to an opensilex instance.
}
