% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_authentication.R
\name{mod_authentification_ui}
\alias{mod_authentification_ui}
\title{authentification UI Function}
\usage{
mod_authentification_ui(
  id,
  default_host = "http://138.102.159.36:8081/rest",
  default_user = "admin@opensilex.org",
  default_password = "admin"
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
The UI part of the module handling the connection to an opensilex instance. Return a token
}
